python-engine/requirements.txt
.\insightface-0.7.3-cp312-cp312-win_amd64.whl
onnxruntime-gpu==1.19.2
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
pydantic==2.8.2
cython==3.0.11
